+++ src/main/java/CertificateOfDeposit.java
@@ -1,6 +1,6 @@
 public class CertificateOfDeposit extends Account {
 
-	public CertificateOfDeposit(double apr, int id, int initialDeposit, String accountType) {
+	public CertificateOfDeposit(double apr, int id, double initialDeposit, String accountType) {
 		super(apr, id, accountType);
 		super.setBalance(initialDeposit);
 	}
